[Category] Text
[Issue] Number is not displayed.

[File] misc/star7.swf

[Steps to reproduce]
1. Play back the file.

[Expected behavior]
A number '3' is displayed at the center of the stage.

[Observed behavior]
No text is displayed.

[Analysis]
- The text is EditText (device text, id=1578, fontId=1577, initial text is '0'.)
- The text is updated to '3'.
- preDraw() and draw() are called.

[Cause]
Font name contains multibyte characters that is not supported by the platform.
There is already code to replace the multibyte font name with the default one.
However, as the code only checks the first character, it is impossible to handle the font name such as '_ゴシック'

[Fix]
Changed to check the entire string of the font name.